Support for quickly importing geometry directly from Blender-internal memory

+++ src/converter/collada.cpp
@@ -485,7 +485,7 @@ void writeGeometry(ColladaContext &ctx, const std::string &prefixName, std::stri
 		stream->close();
 		filename = "meshes/" + filename;
 	} else {
-		ctx.cvt->m_geometryDict.push_back((uint32_t) ctx.cvt->m_geometryFile->getPos());
+		ctx.cvt->m_geometryDict.push_back((uint64_t) ctx.cvt->m_geometryFile->getPos());
 		mesh->serialize(ctx.cvt->m_geometryFile);
 		filename = ctx.cvt->m_geometryFileName.filename().string();
 	}
